+++ NEWS.md
@@ -9,6 +9,17 @@ vegtable 0.1.7
 vegtable 0.1.6
 ==============
 
+<<<<<<< HEAD
+### New Features
+
+### Improvements
+
+* Documentation implemented in `roxygen2`.
+
+### Bug Fixes
+
+* Function `count_taxa()` adapted to the package `taxlist`.
+=======
 ### Improvements
 
 * Documentation implemented in `roxygen2`.
@@ -17,6 +28,7 @@ vegtable 0.1.6
 ### Bug Fixes
 
 * Function `count_taxa()` adapted to the current package `taxlist`.
+>>>>>>> refs/remotes/origin/miguel
 
 vegtable 0.1.5
 ==============